New perform_cleanup() API
perform_cleanup() now only accepts one argument, which is a janitor_spec
data structure which declare the environment variables that are required to
be set for cleanup and also provides a callable function for performing
the cleanup.

Example:

perform_cleanup(janitor_spec={
    'postgres': {
        'required_envvars': ['PTERO_LSF_DB_STRING'],
        'cleanup_action': lambda: PostgresJanitor(
            os.environ['PTERO_LSF_DB_STRING']).clean(),
        },
    })

The purpose of changing the perform_cleanup() API is to allow for better
argument parsing in the Janitor.  Now the purge-backends scripts accept as
arguments the backends which should be cleaned or --all to run the janitor for
all of the backends.

Example usages:
$ PTERO_LSF_DB_STRING=postgres:/dbhost/  purge-backends --postgres

-or-

$ PTERO_LSF_DB_STRING=postgres:/dbhost/  purge-backends --all